Add business formation statistics
